Refactor:MyCurationModal dynamic import 제거
